refactor(search-bar): added data-mapper and transaction layers
