+++ src/components/repos/__mocks__.ts
@@ -1,4 +1,4 @@
-import { Repo } from './repos.store';
+import { Repo } from './repos.types';
 
 export const REPOS: Repo[] = [
     { "name": "tesseract", "description": "Tesseract Open Source OCR Engine (main repository)", "language": "C++", "stars": 50926, "watchers": 50926, "score": 1, "avatar": "https://avatars.githubusercontent.com/u/8401422?v=4", "owner": "tesseract-ocr", "created_at": "2014-08-12T18:04:59Z", "updated_at": "2023-05-14T13:25:30Z", "url": "https://github.com/tesseract-ocr/tesseract", forks: 4342, open_issues: 321 },
