+++ src/components/repos/repos.types.ts
@@ -1,0 +1,18 @@
+export type Repo = {
+    name: string;
+    description: string | null;
+    language: string | null;
+    stars: number;
+    forks: number;
+    open_issues: number;
+    watchers: number;
+    score: number;
+    avatar: string;
+    owner: string;
+    created_at: string;
+    updated_at: string;
+    url: string;
+};
+
+export type SortBy = Omit<keyof Repo, 'description' | 'language' | 'avatar' | 'url'>
+
