+++ src/components/search-bar/search-result.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { RepoLink } from "components/repo-link";
-import { Repo } from 'components/repos/repos.store';
+import { Repo } from 'components/repos/repos.types';
 import { CreatedAt, UpdatedAt } from "components/stats/dates";
 import { ForksCount } from "components/stats/forks-count";
 import { IssuesCount } from "components/stats/issues-count";
