+++ src/utils/get-repo-full-name.ts
@@ -1,3 +1,3 @@
-import { Repo } from 'components/repos/repos.store'
+import { Repo } from 'components/repos/repos.types'
 
 export const getRepoFullName = (repo: Pick<Repo, 'name' | 'owner'>) => `${repo.owner}/${repo.name}`;
